Add GetState to resource monitor
This new method is used to return the state of the resource monitor for
a deployment directly to a client.

This helps with two major areas:
1) `Construct` and `Call` no longer need to be kept up to date with any new
   features added to `Run`. All they need is the resource monitor target
   and can use `GetState` to fetch all other information.
2) We no longer need to find a way to pass all the information from
   `Run` from the runtime host to the runtime SDKs. Instead all we need
   to pass is the (again) the resource monitor target, and on startup
   the SDK can fetch the rest of the information over gRPC. In fact we
   don't even need to add new information to `Run`, as the language
   runtime host can also just use `GetState` to fetch the latest
   information.

A few things need pointing out about this change.

Firstly SDKs that use this _also_ need to have a fallback to using the
old envvar based approch because they might be running against an old
engine version.

Secondly this PR doesn't update any of the SDKs to use this. The most
obvious candidate to update first is YAML as that doesn't actually have
an SDK, just the host and the host _doesn't_ need to be backward
compatible because we always tie host and engine versions together in a
single release.

Thirdly I did consider fixing the "supported version" problem
(https://github.com/pulumi/pulumi-dotnet/pull/88#pullrequestreview-1265714090)
as part of this change. We could have an enum of features and one of the
fields on `MonitorState` could be an array of that enum listing each
supported feature. This is pretty clean and explict but it means in the
mock resource monitors we'd need to explictly list each feature but
we've just made a change to those to instead assume they support all
features and just return `true` for the current `SupportsFeature`
method.